ML-Backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# OCR
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.10

# Vector Store & RAG
faiss-cpu==1.8.0.post1

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2

# CORS
python-jose==3.3.0

# Google Generative AI
google-generativeai==0.8.5